refactor: rename star-con to star-icon
