+++ src/assets/icons/star-icon/star-icon.scss
