+++ src/assets/icons/star-icon/star-icon.tsx
